requirements.txt
asyncio
aiohttp
sanic
python-dotenv